Revert "fix bug in graphael line chart when all y-axis data is the same value.  (as described here: http://groups.google.com/group/raphaeljs/browse_thread/thread/eb2bda8e179bac07/ad8c679d02ede394?lnk=gst&q=+line+chart#ad8c679d02ede394)"
This reverts commit 38dd47361917b4e3b3e53155eb3f46674bb48074.

Conflicts:

	g.line-min.js
	g.line.js

Reverted fixes of issue 9 by Huned.

The addition of a small fraction to the maximum y-value was causing the
values of the y-axis labels to not match the values of plotted points.
This was especially noticeable with small ranges.  (For instance, points
ranging from 0 to 1 would have an axis with range 0 to 1.5, but the
points with value of 1 would be at the same height as the 1.5 label.)